Make sure to use long for connection handle, and generalised context storage
svn path=/trunk/eXist/; revision=14904


Former-commit-id: a34b1b39d726a1cff3a4b290b1c83196d0988042

+++ extensions/modules/src/org/exist/xquery/modules/sql/ExecuteFunction.java
@@ -85,7 +85,7 @@ public class ExecuteFunction extends BasicFunction
             new QName( "execute", SQLModule.NAMESPACE_URI, SQLModule.PREFIX ),
             "Executes a SQL statement against a SQL db using the connection indicated by the connection handle.",
             new SequenceType[] {
-                new FunctionParameterSequenceType( "connection-handle", Type.INTEGER, Cardinality.EXACTLY_ONE, "The connection handle" ),
+                new FunctionParameterSequenceType( "connection-handle", Type.LONG, Cardinality.EXACTLY_ONE, "The connection handle" ),
                 new FunctionParameterSequenceType( "sql-statement", Type.STRING, Cardinality.EXACTLY_ONE, "The SQL statement" ),
                 new FunctionParameterSequenceType( "make-node-from-column-name", Type.BOOLEAN, Cardinality.EXACTLY_ONE, "The flag that indicates whether the xml nodes should be formed from the column names (in this mode a space in a Column Name will be replaced by an underscore!)" )
             },
@@ -95,7 +95,7 @@ public class ExecuteFunction extends BasicFunction
             new QName( "execute", SQLModule.NAMESPACE_URI, SQLModule.PREFIX ),
             "Executes a prepared SQL statement against a SQL db.",
             new SequenceType[] {
-                new FunctionParameterSequenceType( "connection-handle", Type.INTEGER, Cardinality.EXACTLY_ONE, "The connection handle" ),
+                new FunctionParameterSequenceType( "connection-handle", Type.LONG, Cardinality.EXACTLY_ONE, "The connection handle" ),
                 new FunctionParameterSequenceType( "statement-handle", Type.INTEGER, Cardinality.EXACTLY_ONE, "The prepared statement handle"),
                 new FunctionParameterSequenceType( "parameters", Type.ELEMENT, Cardinality.ZERO_OR_ONE, "Parameters for the prepared statement. e.g. <sql:parameters><sql:param sql:type=\"varchar\">value</sql:param></sql:parameters>"),
                 new FunctionParameterSequenceType( "make-node-from-column-name", Type.BOOLEAN, Cardinality.EXACTLY_ONE, "The flag that indicates whether the xml nodes should be formed from the column names (in this mode a space in a Column Name will be replaced by an underscore!)" )
